requirements.txt
langgraph
langchain-anthropic
langchain-community
tavily-python
pypdf
python-dotenv
requests
rich